pyproject.toml
[project]
name = "jmtech-ecommerce"
version = "0.1.0"
description = "Backend for JMTech E-commerce"
authors = [
    {name = "Antigravity"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "fastapi (>=0.122.0,<0.123.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "sqlmodel (>=0.0.27,<0.0.28)",
    "alembic (>=1.17.2,<2.0.0)",
    "psycopg[binary] (>=3.2.13,<4.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
